Fix some down conversions from double to float so they are explicit

+++ src/cache.h
@@ -279,9 +279,9 @@ __attribute__((always_inline)) INLINE void cache_read_particles_subset(
      * interact. We use values of the same magnitude (but negative!) as the real
      * particles to avoid overflow problems. */
     const double max_dx = ci->dx_max_part;
-    const float pos_padded[3] = {-(2. * ci->width[0] + max_dx),
-                                 -(2. * ci->width[1] + max_dx),
-                                 -(2. * ci->width[2] + max_dx)};
+    const float pos_padded[3] = {(float)-(2. * ci->width[0] + max_dx),
+                                 (float)-(2. * ci->width[1] + max_dx),
+                                 (float)-(2. * ci->width[2] + max_dx)};
     const float h_padded = ci->parts[0].h;
 
     for (int i = *last_pi; i < *last_pi + VEC_SIZE; i++) {
@@ -327,9 +327,9 @@ __attribute__((always_inline)) INLINE void cache_read_particles_subset(
      * interact. We use values of the same magnitude (but negative!) as the real
      * particles to avoid overflow problems. */
     const double max_dx = ci->dx_max_part;
-    const float pos_padded[3] = {-(2. * ci->width[0] + max_dx),
-                                 -(2. * ci->width[1] + max_dx),
-                                 -(2. * ci->width[2] + max_dx)};
+    const float pos_padded[3] = {(float)-(2. * ci->width[0] + max_dx),
+                                 (float)-(2. * ci->width[1] + max_dx),
+                                 (float)-(2. * ci->width[2] + max_dx)};
     const float h_padded = ci->parts[0].h;
 
     for (int i = ci->count - *first_pi; i < ci->count - *first_pi + VEC_SIZE;
@@ -530,9 +530,9 @@ __attribute__((always_inline)) INLINE void cache_read_two_partial_cells_sorted(
    * interact. We use values of the same magnitude (but negative!) as the real
    * particles to avoid overflow problems. */
   const double max_dx = max(ci->dx_max_part, cj->dx_max_part);
-  const float pos_padded[3] = {-(2. * ci->width[0] + max_dx),
-                               -(2. * ci->width[1] + max_dx),
-                               -(2. * ci->width[2] + max_dx)};
+  const float pos_padded[3] = {(float)-(2. * ci->width[0] + max_dx),
+                               (float)-(2. * ci->width[1] + max_dx),
+                               (float)-(2. * ci->width[2] + max_dx)};
   const float h_padded = ci->parts[0].h;
 
   for (int i = ci->count - first_pi_align;
@@ -606,9 +606,9 @@ __attribute__((always_inline)) INLINE void cache_read_two_partial_cells_sorted(
   /* Pad cache with fake particles that exist outside the cell so will not
    * interact. We use values of the same magnitude (but negative!) as the real
    * particles to avoid overflow problems. */
-  const float pos_padded_j[3] = {-(2. * cj->width[0] + max_dx),
-                                 -(2. * cj->width[1] + max_dx),
-                                 -(2. * cj->width[2] + max_dx)};
+  const float pos_padded_j[3] = {(float)-(2. * cj->width[0] + max_dx),
+                                 (float)-(2. * cj->width[1] + max_dx),
+                                 (float)-(2. * cj->width[2] + max_dx)};
   const float h_padded_j = cj->parts[0].h;
 
   for (int i = last_pj_align + 1; i < last_pj_align + 1 + VEC_SIZE; i++) {
@@ -724,9 +724,9 @@ cache_read_two_partial_cells_sorted_force(
    * interact. We use values of the same magnitude (but negative!) as the real
    * particles to avoid overflow problems. */
   const double max_dx = max(ci->dx_max_part, cj->dx_max_part);
-  const float pos_padded[3] = {-(2. * ci->width[0] + max_dx),
-                               -(2. * ci->width[1] + max_dx),
-                               -(2. * ci->width[2] + max_dx)};
+  const float pos_padded[3] = {(float)-(2. * ci->width[0] + max_dx),
+                               (float)-(2. * ci->width[1] + max_dx),
+                               (float)-(2. * ci->width[2] + max_dx)};
   const float h_padded = ci->parts[0].h;
 
   for (int i = ci->count - first_pi_align;
@@ -788,9 +788,9 @@ cache_read_two_partial_cells_sorted_force(
   /* Pad cache with fake particles that exist outside the cell so will not
    * interact. We use values of the same magnitude (but negative!) as the real
    * particles to avoid overflow problems. */
-  const float pos_padded_j[3] = {-(2. * cj->width[0] + max_dx),
-                                 -(2. * cj->width[1] + max_dx),
-                                 -(2. * cj->width[2] + max_dx)};
+  const float pos_padded_j[3] = {(float)-(2. * cj->width[0] + max_dx),
+                                 (float)-(2. * cj->width[1] + max_dx),
+                                 (float)-(2. * cj->width[2] + max_dx)};
   const float h_padded_j = cj->parts[0].h;
 
   for (int i = last_pj_align + 1; i < last_pj_align + 1 + VEC_SIZE; i++) {
